+++ apps/ui-dev/pages/ui-component/data-editor.jsx
@@ -1,0 +1,55 @@
+import AddFilled from "@elf-framework/icon/esm/AddFilled";
+import { start, SUBSCRIBE, UIElement, useSubscribe } from "@elf-framework/sapa";
+import { DataEditor, View } from "@elf-framework/ui";
+import "@elf-framework/ui/style.css";
+
+import { Layout } from "../../component/Layout";
+
+class Sample extends UIElement {
+  template() {
+    return <div>sample</div>;
+  }
+
+  [SUBSCRIBE("config:select.change")]() {
+    console.log("config:select.change");
+  }
+}
+
+function App() {
+  useSubscribe("refreshIcon", () => {
+    console.log("refreshIcon");
+  });
+
+  return (
+    <Layout>
+      <View width={300}>
+        <DataEditor
+          data={() => ({
+            title: "Data Editor",
+          })}
+          items={(data, root) => {
+            return [
+              {
+                type: "button",
+                title: "Add",
+                icon: <AddFilled />,
+                onClick: () => {
+                  this.emit("config:select.change");
+                  console.log("Button clicked");
+                },
+              },
+              {
+                type: "color",
+                key: "color",
+                value: "#ff0000",
+              },
+            ];
+          }}
+        />
+        <Sample />
+      </View>
+    </Layout>
+  );
+}
+
+start(App);