+++ apps/ui-dev/vite.dev.config.js
@@ -1,5 +1,20 @@
+import * as glob from "glob";
 import { defineConfig } from "vite";
 import sapa from "vite-plugin-sapa";
+
+import path from "path";
+
+const entries = {};
+const files = glob.sync("pages/**/*.html", {
+  dot: true,
+  // node_modules 은 검색대상에서 제외
+  ignore: ["node_modules/**"],
+});
+
+files.forEach((it) => {
+  const file = path.basename(it, ".html");
+  entries[file] = path.resolve(__dirname, it);
+});
 
 export default defineConfig({
   server: {
@@ -17,12 +32,20 @@ export default defineConfig({
     jsxFragment: "FragmentInstance",
     jsxInject: `import { createElementJsx, FragmentInstance } from "@elf-framework/sapa"`,
   },
+  build: {
+    rollupOptions: {
+      input: {
+        ui: path.resolve(__dirname, "index.html"),
+        ...entries,
+      },
+    },
+  },
   optimizeDeps: {
     exclude: [
+      "@elf-framework/design-tokens",
       "@elf-framework/sapa",
       "@elf-framework/ui",
       "@elf-framework/sapa-router",
-      "@elf-framework/icon",
       "@elf-framework/design-system",
     ],
   },