+++ packages/ui-react/src/index.ts
@@ -1,5 +1,0 @@
-import "@elf-framework/design-system/style.css";
-
-export * from "./components/button";
-export * from "./components/checkbox";
-export * from "./components/color/ColorView";