+++ packages/vite-plugin-sapa/CHANGELOG.md
@@ -1,5 +1,17 @@
 # vite-plugin-sapa
 
+## 2.0.33
+
+### Patch Changes
+
+- modify sapa refresh logic
+
+## 2.0.32
+
+### Patch Changes
+
+- fixed sapa refresh logic
+
 ## 1.1.1
 
 ### Patch Changes